Issue/366: Extracted roll formula resolver code to its own class
* This should make it easier to re-implement this feature in other places, should the need arise.
